+++ tests/BusinessLogic/Tasks/UpdateShippingServicesTaskTest.php
@@ -366,6 +366,7 @@ class UpdateShippingServicesTaskTest extends BaseSyncTest
                 new HttpResponse(200, array(), '[]'),
                 new HttpResponse(200, array(), '[]'),
                 new HttpResponse(200, array(), '[]'),
+                new HttpResponse(200, array(), '[]'),
                 new HttpResponse(200, array(), '[]')
             )
         );
@@ -435,6 +436,7 @@ class UpdateShippingServicesTaskTest extends BaseSyncTest
             new HttpResponse(200, array(), $this->getDemoServiceDeliveryDetails('IT-CL')),
             new HttpResponse(200, array(), $this->getDemoServiceDeliveryDetails('IT-CZ')),
             new HttpResponse(200, array(), $this->getDemoServiceDeliveryDetails('IT-SE')),
+            new HttpResponse(200, array(), '[]'),
             new HttpResponse(200, array(), '[]'),
             new HttpResponse(200, array(), '[]'),
             new HttpResponse(200, array(), '[]'),