+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [3.7.4](https://github.com/packlink-dev/ecommerce_module_core/compare/v3.7.3...v3.7.4) - 2026-02-06
+### Changed
+- Fix shipping service special tag check
+
+## [3.7.3](https://github.com/packlink-dev/ecommerce_module_core/compare/v3.7.2...v3.7.3) - 2026-02-04
+### Changed
+- Add weight conversions
+- Fixed service sync issue for AU, CA, and US
+
+## [3.7.2](https://github.com/packlink-dev/ecommerce_module_core/compare/v3.7.1...v3.7.2) - 2026-01-14
+### Changed
+- Add postal code for Lithuania
+- 
 ## [3.7.1](https://github.com/packlink-dev/ecommerce_module_core/compare/v3.7.0...v3.7.1) - 2025-10-20
 ### Changed
 - Add postal code for South Korea